+++ cmd/query-tee/proxy_test.go
@@ -1,0 +1,191 @@
+package main
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"strconv"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/go-kit/kit/log"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func Test_NewProxy(t *testing.T) {
+	cfg := Config{}
+
+	p, err := NewProxy(cfg, log.NewNopLogger(), nil)
+	assert.Equal(t, errMinBackends, err)
+	assert.Nil(t, p)
+}
+
+func Test_Proxy_RequestsForwarding(t *testing.T) {
+	const (
+		querySingleMetric1 = `{"status":"success","data":{"resultType":"vector","result":[{"metric":{"__name__":"cortex_build_info"},"value":[1583320883,"1"]}]}}`
+		querySingleMetric2 = `{"status":"success","data":{"resultType":"vector","result":[{"metric":{"__name__":"cortex_build_info"},"value":[1583320883,"2"]}]}}`
+		queryEmpty         = `{"status":"success","data":{"resultType":"vector","result":[]}}`
+	)
+
+	type mockedBackend struct {
+		pathPrefix string
+		handler    http.HandlerFunc
+	}
+
+	tests := map[string]struct {
+		backends            []mockedBackend
+		preferredBackendIdx int
+		expectedStatus      int
+		expectedRes         string
+	}{
+		"one backend returning 2xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric1)},
+			},
+			expectedStatus: 200,
+			expectedRes:    querySingleMetric1,
+		},
+		"one backend returning 5xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 500, "")},
+			},
+			expectedStatus: 500,
+			expectedRes:    "",
+		},
+		"two backends without path prefix": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric1)},
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric2)},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      200,
+			expectedRes:         querySingleMetric1,
+		},
+		"two backends with the same path prefix": {
+			backends: []mockedBackend{
+				{
+					pathPrefix: "/api/prom",
+					handler:    mockQueryResponse("/api/prom/api/v1/query", 200, querySingleMetric1),
+				},
+				{
+					pathPrefix: "/api/prom",
+					handler:    mockQueryResponse("/api/prom/api/v1/query", 200, querySingleMetric2),
+				},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      200,
+			expectedRes:         querySingleMetric1,
+		},
+		"two backends with different path prefix": {
+			backends: []mockedBackend{
+				{
+					pathPrefix: "/prefix-1",
+					handler:    mockQueryResponse("/prefix-1/api/v1/query", 200, querySingleMetric1),
+				},
+				{
+					pathPrefix: "/prefix-2",
+					handler:    mockQueryResponse("/prefix-2/api/v1/query", 200, querySingleMetric2),
+				},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      200,
+			expectedRes:         querySingleMetric1,
+		},
+		"preferred backend returns 4xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 400, "")},
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric1)},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      400,
+			expectedRes:         "",
+		},
+		"preferred backend returns 5xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 500, "")},
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric1)},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      200,
+			expectedRes:         querySingleMetric1,
+		},
+		"non-preferred backend returns 5xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 200, querySingleMetric1)},
+				{handler: mockQueryResponse("/api/v1/query", 500, "")},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      200,
+			expectedRes:         querySingleMetric1,
+		},
+		"all backends returns 5xx": {
+			backends: []mockedBackend{
+				{handler: mockQueryResponse("/api/v1/query", 500, "")},
+				{handler: mockQueryResponse("/api/v1/query", 500, "")},
+			},
+			preferredBackendIdx: 0,
+			expectedStatus:      500,
+			expectedRes:         "",
+		},
+	}
+
+	for testName, testData := range tests {
+		t.Run(testName, func(t *testing.T) {
+			backendURLs := []string{}
+
+			// Start backend servers.
+			for _, b := range testData.backends {
+				s := httptest.NewServer(b.handler)
+				defer s.Close()
+
+				backendURLs = append(backendURLs, s.URL+b.pathPrefix)
+			}
+
+			// Start the proxy.
+			cfg := Config{
+				BackendEndpoints:   strings.Join(backendURLs, ","),
+				PreferredBackend:   strconv.Itoa(testData.preferredBackendIdx),
+				ServerServicePort:  0,
+				ServerMetricsPort:  0,
+				BackendReadTimeout: time.Second,
+			}
+
+			p, err := NewProxy(cfg, log.NewNopLogger(), nil)
+			require.NoError(t, err)
+			require.NotNil(t, p)
+			defer p.Stop() //nolint:errcheck
+
+			require.NoError(t, p.Start())
+
+			// Send a query request to the proxy.
+			res, err := http.Get(fmt.Sprintf("http://%s/api/v1/query", p.Endpoint()))
+			require.NoError(t, err)
+
+			defer res.Body.Close()
+			body, err := ioutil.ReadAll(res.Body)
+			require.NoError(t, err)
+
+			assert.Equal(t, testData.expectedStatus, res.StatusCode)
+			assert.Equal(t, testData.expectedRes, string(body))
+		})
+	}
+}
+
+func mockQueryResponse(path string, status int, res string) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		// Ensure the path is the expected one.
+		if r.URL.Path != path {
+			w.WriteHeader(http.StatusNotFound)
+			return
+		}
+
+		// Send back the mocked response.
+		w.WriteHeader(status)
+		if status == http.StatusOK {
+			_, _ = w.Write([]byte(res))
+		}
+	}
+}